[AUA] Revise buildConfig default processor
Instead of attempting to inspect the source code of the project, which
for any kind of correctness required generated code (and which was not
in any case fully correct given that we only sync a single Variant of
the project), preserve the pre-8.0.0-alpha09 behaviour by inserting a
directive into the gradle.properties file.

This also probably addresses some exceptions, by not requiring indexes
during initialization and management of component processors.

In the interests of cherry-picking to the release branch, this change
is minimal; it does not replace the existing heavy tests with lighter
tests (now that we no longer require generating sources and
introspecting models).  These cleanups will be addressed in a
follow-up change.

Bug: 260847161
Test: existing tests adjusted
Change-Id: I27b63868e4fd3868372656bc231a64ae0523fe4f
(cherry picked from commit 38ee2897059ce19c2cd553fe3cacb85273d45aaa)