process name monitor: Use legacy shell
The number of object allocations when using shell-v2 is a lot
higher than when using shell-v1. This is especially pronnounced
in the case of the ProcessNameClientMonitor since it invokes
a `ps` command every 2 seconds.

This CL switches ProcessNameClientMonitor to use the legacy
shell, and can be reverted when we can confirm that either this
is not a significant issue, or if we fix the shell-v2 behavior.

Bug: 262179063
Test: verified that both shell v2 and v1 return similar outputs at one
API level, an opt out is available via the studio flag.

Change-Id: I22ae54dd22083492ce1ab2c9def770d2953468a1