[Ps] Provide a refactoring for buildConfig
If the project is using a version of AGP where the default is true,
adjust module build files so that we can insert a property setting the
default to false.

If the project is using a version of AGP where the default is false,
perform any necessary module build file modifications and then remove
the property.

Bug: 262240431
Bug: 289219116
Test: manual, added
Change-Id: I1440eaa125f2e2ab97d954d20391260badb72310