Search only for XDebuggerTree in DebuggerTest
DebuggerTest had been searching for an XDebuggerFramesList, which only
used to work due to a race condition consistently working in our favor:
when the debugger frames are loading, the Swing hierarchy shows an
XDebuggerFramesList$XDebuggerFrameListRenderer while the "Loading..."
message is on-screen. After that, XDebuggerFramesList won't be found by
name, but it DOES exist as an anonymous class: XFramesView$2.

The search was pointless anyway since the presence of XDebuggerTree
alone is enough to verify that the breakpoint worked.

Bug: 245983885
Test: DebuggerTest
Change-Id: I7fb44d2ce747cd2079df8c1d6693de49df3ac913