263689310: Support quickfix preview
This CL updates the various lint-related quickfixes in the IDE to
properly support intention action previews, which are on by default in
the latest IntelliJ EAPs.

First, this was triggering some threading exceptions (such as
   263689310: "Access is allowed from write thread only" in
   SuppressLintQuickFix.handleKotlin
) because the preview editor support invokes these actions outside of
write actions even though the actions are marked startInWriteThread (see
IntentionAction.generatePreview).

However, even after fixing the exceptions, many actions didn't support
preview (generally because they relied on looking up the editor document
to change via the document manager, and that isn't valid for preview
files). This CL also fixes that.

Fixes: 263689310
Test: Many existing tests for lint quickfixes in AndroidLintTest and
      LintIdeTest (there are tests for the quickfixes; none of them
      relied on preparePsiElementForWrite to do something useful)
Change-Id: Ic90f07e30ad206ad55d51f73aaf6ff0d47bced96