Make the test suite register IconManager first
RenderLogicTests fail without initialziing the IconManager first in the
test suite. This is due to the degree.svg icon being loaded as a
DummyIcon before the RenderLogicTests use the active ImageIcon for
rendering. Since the tests suite is dependent on icon images being
rendered, register the IconManager first to prevent any concurrency
issues between the tests.

Bug: 201230339
Change-Id: If4c8f13fd8268f75c8e580832bdc161a9365efb0