Adapt to generalized start/stop trace commands
A change on the base side converted the start and
stop cpu trace commands and their handlers to a more
generic start and stop trace command. This allows us
to encapsulate the memory and other types of traces
into one command.

This change adapts the studio-side code to the base-side
proto changes.

Bug: 263172101
Test: updated
Change-Id: I41a47ebb60ab4a70d2ef4a3098b00bcee46a3b93