Fix TransitiveTestRClassesTest.testLibTestResources
For whatever reason, the IDEA 223.6160.11 update caused VirtualFile
updates to be queued slightly later than they were before. In this test,
a call to check highlighting obtained a lock on the EDT before the
updates finished, but of course the updates need to happen on the EDT as
well. This leads to a logical deadlock where the `checkHighlighting`
method loops its maximum 1000 times and fails.

It looks like we've run into this before, because there's an existing
utility method in this set of tests to handle that case.

Fixes: 264602834
Test: TestRClassesTest
Change-Id: Ia94569ba84f1e741006c70470c62c4445d638df2