Fix flakey tests
testStopPollingSelectedDevice had a race condition. Sometimes device2
was connected before device1 was set as the selectedDevice. This
incorrectly caused foreground process detection to start on device2.
The problem is solved by adding a sync point on the change of
selectedDevice.

This change also runs every `channel#send` operation in a separate
coroutine.
Because send and receive are blocking, this makes sure that if we call
send before calling receive we don't block the main coroutine, which
would prevent receive from ever being called.

Fixes: 264825631
Test: ForegroundProcessDetectionTest
Change-Id: I7549779590b3fb341104609d5c5287df2b05f774