Make the power profiler feature flag dynamic
The feature flag for displaying the power rail and battery
counter tracks was made to be an enum/integer to be more dynamic.
This now allows us to disable (PowerProfilerDisplayMode.HIDE) the
viewing of the power and battery tracks, but allows us to add different
visualizations via non-zero positive integers mapping to a different
visualization query.

Bug: 254116554
Test: updated
Change-Id: I3a656229b12ee030d0299c063781d20e3ad6a098