[BA] Fix exception on UI after switching themes
We recreate UI components on UI theme change to make sure new values
are propagated to all elements properly. We keep the existing model of
the pages to not reset the state of the UI.
The problem was that we did not clear UI listeners on the model so
when user starts browsing BA UI, then switches UI theme and continue
browsing BA UI, both disposed and new UI elements are notified by the
model.

Fix by making sure we remove the listeners on UI disposal.

Fixes: 259191641
Test: Add testcase to BuildAnalyzerComboBoxViewTest
Change-Id: I2cf46c3cfea75effe5b4f819ba4941ef207810a3
(cherry picked from commit 835bbec7d35ab822011f2feee37290f648fc3a1d)