Populate light R class for modules with real values
This change ensures that, if available, the real values for the ids are
used instead of fake generated ones.
Light R classes are used by the Live Edit compiler that re-uses the
analysis from the editor. If fake values are used, and the values are
static final, they will be inlined by the compiler in the Live Edit
compiled code leading to unresolved resources.

Removed the field initializers in the light R class to avoid the fake
values being displayed in auto-complete.

Fixes: 265636305
Test: FastPreviewGradleTest
Change-Id: Ia1a5ea0c0be0bf29478d99559e4cd06e80ae832d