Allow current rail to display neg vals
The current UI infrastructure does not allow
for negative values to be displayed in the tracks.

This change allows the display of the negative values
in the track UI. It also works around the aforementioned
infrastructure to allow for a label of 0 to show negative
and positive values relative values.

Bug: 254116554
Test: added
Change-Id: I80f85b6c8e7567b9c452d8206264b88528257bb0