Refactor RecordingManager.java
This change makes RecordingManager a Light Service instead of a
static singleton. Because it is also a Disposable, it means we
can attach CoroutineScopes for related classes (namely the
JfrReportGenerators) to it instead of just attaching to the
JobScheduler or a new thread pool.

Note that instead of using "synchronized" which may not play well
with coroutines, code accessing formerly guarded members is
guarded by a Kotlin coroutines-aware mutex.

Bug: N/A
Test: JfrReportGeneratorTest.kt
Change-Id: If9b2947a516b2e0758198760b7e35fd7b13c2e39