Do not refresh in ApkModuleNode constructor
There shouldn't be any need to refresh inside a tree node constructor,
the code that modified the APK / dex files should be responsible for
triggering the refresh, if necessary.

And, if that refresh detects a change to the file, I believe that
should trigger the file tree/children to be updated automatically.

Bug: 264160639
Test: Manual, enable/disable multidex and reload APK
Change-Id: I8d4869a6b27e83f5eebee406466c16e08ac92b56
(cherry picked from commit 33a02cfb315725739ab293af4f77026d1442028e)