Switch to using transitive dependencies when available
The V2 models allow us to use the transitive dependencies in the PSD
rather than relying on the parsing of POM files which can be fragile.

This change switches to using these when available, in an upcoming
change we will not longer be fetching the POM file paths from Gradle.

For older AGP versions POM files should still be avaiable and we fall
back to these when the transitive dependencies are not available.

Bug: 197636221
Test: current test now covers new behaviour, added a test to cover for
old agps

Change-Id: Icc4a4a4e301c17f891322360d65619769934c8cc