Fix not properly disposed objects in test
Apparently disposable object in this test was registered incorrectly
with DisposableRule (problem 1) and disposable rule does not check
itself for leaked disposables (problem 2), this check happens as part
of android project rules.
This was discovered only accidentally by running another test with
project rule after this one.
Fix both problems locally now, need to figure out how to avoid it in the
future.

Bug: N/A
Test: Fixing test
Change-Id: Ic88b06e53cdf536bd436a7e3a6d3e85beb71eb77