Add toggle to experimental setting for target device api optimization
Currently when running Android app from IDE, we end up injecting the target device API to the Gradle process so that it optimizes various tasks and transforms. On the flip side, some devs have been complaining that they would rather not have this feature enabled by default (so they would prefer not to inject the API level) as in some specific cases it may lead to cache misses. From our perspective, we probably want to keep this feature enabled (until we have indisputable proof that is it actually harmful) but also allow however wants to disable the injection of device api level.

Bug: 164145066
Test: updated relevant test classes
Change-Id: I4e3a4d8825820e287fd049d66b18522f77275abc