Add downloads info presentation for sync and build
This change adds visual presentation of file download requests
to Sync and Build output.
This new view is installed as a custom ExecutionConsole via a
PresentableBuildEvent. This allows us to install any custom UI that
will be shown on the corresponding node selection.
The event is sent right on the start of build to be presented as the
first item in the build output tree.
The custom UI consists of two tables, first shows per repository summary
(including total all together summary), second table presents requests
list, filtered by repo selection in the first table if any.
On issuing this event (installing the UI) we keep reference to the
downloads data model. On the other end we subscribe to download events
from gradle and feed this data to the model which updates UI live.

Receiving and processing these events was already done in Build Analyzer
for builds and this applies some refactorings to this code to reuse this
code for the new purpose as well. Receiving these events on Sync and
reporting it on stats was already done a bit earlier so this change
extends this infrastructure to report downloads on UI in a way described
above.

Bug: 265924450
Test: Added unit tests, integration test to follow in a separate change.
Change-Id: I2666084f816e8c2d3f57bad5646d52f7929a3854