Update DownloadsAnalyzerTest
As we request source/javadoc/sample when building the android model,
there are some additional http requests. The duplicates http requests
are due to the fact that we are currently requesting them from ide as
well. But once we stop requesting them from IDE(use the one from android
model), those duplicates should be gone. This is verified by commenting
out getAdditionalClassifierArtifactsModel involcation in
SyncProjectActionWorker.

Bug: 197636221
Test: existing
Change-Id: Iaea502d622ba25049d513b460ca887e8534af048