Add new extensions for Dagger using index
It turns out to be more straightforward to provide entirely separate
implementations of these three extensions using the new index than to
try to incorporate the logic into the existing classes. I've named these
classes with the "V2" suffix for the time being; after shipping and
deleting the old code, that will be removed.

There are currently no automated tests, since this is a work in progress
and part of a larger change. The existing extensions don't lend
themselves to an easy unit test with this flag, and the new extensions
don't have any real logic that would allow them to be verified yet.
(Both flag on and flag off operate the same.) Appropriate testing will
be added as the implementations are completed.

Bug: 265846405
Test: Manually verified
Change-Id: Idd25b4695b601a4a59b89e8f760aae0b393fe2d1