New PsiCodeFileChangeDetectorService
This new service relies on observing PSI tree events instead of
generating hashes for the code. This is cheaper as it will only listen
for the first relevant change of a file and that will be enough to
detect the change. The previous implementation relied on calculating a
hash for the entire PSI tree of a file.

Additionally, this also allows for more precise control of which files
are out of date for a given project. Using this information, Fast
Preview is able to better detect when it needs to run and also detect
the cases where it could not get the project to be fully up to date (for
example, when modifying Java files).
When modifying Java files, Fast Preview will continue compiling the Java
bits but will correctly mark the project as being potentially out of
date.
So now, even with Fast Preview enabled, there are cases where we might
report the preview as out of date.

GlancePreview still uses some of the old code while we migrate to the
new mode.

Test: PsiCodeFileChangeDetectorServiceTest
Bug: 199384558
Change-Id: I5b173c5050cb74c8a3140a3506fa971d8fdea5bf