[Dsl] Relax requirements on ReferenceTo contexts
The context really only needs to implement one method, not everything
in GradleDslElementModel.  Separate out that interface from the other,
in order that we can use a GradleBuildModel as a context.

Why do we need to use a GradleBuildModel as a context?  I'm so glad
you asked!  We don't have an explicit block representation of the
plugins { ... } Dsl block in build files, because the GradleBuildModel
attempts to smooth over the many ways of applying plugins to a
project.  But this means that when attempting to add a plugin by
reference (e.g. to a Version Catalog entry) we do not have a block
model to use as the context for the reference.  Allowing the build
file itself to be used as the context is, if not wholly correct, at
least more correct than any other context we have available.

Bug: 267607937
Test: in next CL
Change-Id: I9591fe3334ee5b30b44b58b726e73ea9b3897bc7