Prevent concurrent captures in profiler
Before, when the user would click to stop capturing,
the stop runnable would run, and then immediately after
setFinished() would be called to set isRecording to false.
Thus, even in the case where a stop command execution would
wait and/or timeout, it acted as if the recording state was
terminated. This allowed for new captures to be started even
when a prior one was not terminated yet.

This change extracts the call to setFinished() to only be
called in the trace stop status event handler. In this way,
isRecording is not set to false until we have confirmation that
the stop command has reported back (whether successfuly or in failure).

The heap dump and java/kotlin allocation recording options do not
need to call setFinished() as they originally did not call it. This
is because they do not utilize the "Stop" button located under the
recording options to stop the capture. The "Stop" button was the only
way to call the RecordingOptionsModel::stop method, and this stop()
method was the only method that called setFinished().

Fixes: 264953386
Test: updated + added
Change-Id: Ia77f8d7f991612458ddccc392f80d46e72730924