Add logic to resolve PsiElements for Dagger
Each DaggerConcept is responsible for analyzing PsiElements at runtime
to determine if they represent the construct it handles. Additionally,
it owns the related task of "inflating" an index value into a
PsiElement. (This is related since after the index value's element is
found, it has to be checked to figure out if it's really the Dagger
element we think it is.)

This change also introduces the concept of a `DaggerElement`, which will
be what is exposed to consumers outside the index package. This is a
wrapper around a PsiElement, the JVM type it interacts with, and the
"type" of Dagger element (Provider, Consumer, etc.).

Test: Included
Bug: 265846405
Change-Id: I395523a1a8199f3baaaae2c94b75d1afbbfb26ec