Create new ComboBox on Gradle settings to modify java.home property
To be able to modify the java.home from config.properties a dedicated
ComboBox was added bellow the existing GradleJdkComboBox. This will
be enabled only when users have selected the #GRADLE_LOCAL_JAVA_HOME
gradleJvm for their current project gradle root.

In addition, this component provides suggested valid JDKs from the
jdk.table plus the embedded one, but also allows the possibility to
browse. On top of that the provided path is validated before to
apply changes.

Component follows the Intellij guidelines:
https://jetbrains.design/intellij/controls/built_in_button/#browse

Bug: 214428183
Test: GradleJdkPathComboBoxTest
Test: GradleJdkPathComboBoxBuilderTest
Change-Id: I5b05f4acc822fdf30164dd4ccf32189fa3a0bf4d