Introduce configurable Default JDK and expose it on Gradle settings
A new AndroidDefaultGradleSystemSettingsControlBuilder has been
created that allows us to display the desired view instead of
IdeaGradleSystemSettingsControlBuilder for default projects.
In addition, this new controlBuilder is reusing the existing
GradleJdkPathEditComboBox that makes jdk path selection easier
to the user by suggesting and validating the selected path.

The default JDK represents the jdk path that will first considered,
if defined, to configure the new projects JDK otherwise the embedded
JDK will be used.

In addition this will be stored using PropertiesComponent, that
will persist on disk the current selection, similar to sdk.dir.

Bug: 214428183
Test: GradleConfigManagerTest
Test: ImportProjectGradleSyncJdkIntegrationTest
Test: GradleDefaultJdkPathStoreTest
Change-Id: I0fec941b68b9226a0122545277b14e120b8a0366