[Dsl] Adjust SdkOrPreviewTransform
Only use the -SdkPreview form of the setter if we are either: setting
the Sdk to a literal String beginning "android-" (in which case,
remove "android-" from the value) or setting it to a value (possibly
by reference) to a String *not* beginning "android-".

This addresses the immediate problem of the Project Structure Dialog
setting compileSdkPreview to an unacceptable string for AGP versions
4.1.0 or later.  However, we need to do more work (in a subsequent CL)
in order to have the model parsed data be consistent between the
pre-4.1.0, as demonstrated by the commented-out line in the added
test.

Bug: 267259672
Test: added
Change-Id: If9c76ad4b1e79027eb1d91e761c9fc4c43a7bf4f