Allow to add custom logic for the changes in SceneLayer.showOnHover
In SceneLayer, the changes of 'showOnHover' only depend on whether
the hover interaction happens over its corresponding SceneView or
not. However, in some scenarios we may want to have the ability to
add some extra logic to this decision. For example for Compose
Preview, we want to only show the hover state for selected previews.

This change adds a showOnHoverFilter to the SceneLayer, that screen
view providers could later use to dinamically influence the decisions
around showOnHover.

Bug: 232719077
Test: added
Change-Id: I56fa18cecfc260569c3e2e89dfe3454b0234beb0