Appropriately handle AndroidSyncException with sync issue reporting
When a gradle sync fail with a fatal error that result in no computed variant, ProjectImportAction fails with AndroidSyncException. This change adds the sync issues to the exception and report them through SyncIssueReporter. These issues are added to the DataNode and reported to the users along with the thrown exception. Throwing the sync exception fails the requested sync and the build output window is opened automatically to show users the sync issues with a link to the build file where the error originated.

Bug: 267726572
Test: Existing Tests
Change-Id: Ia0d7175893bbbb0c64bfd87362056cd45db0a1fc