Fix sdk patcher
The sdk patcher is itself an sdk component, so if we need to change the
patch format we can do so and update the patcher independent of studio,
and also without having to rebuild past patches.

The patcher jar includes almost all non-jre classes needed, but since
class resolution is by default done by the parent loader first, we were
incorrectly picking up log4j from studio rather than the version bundled
with the patcher. When studio updated log4j, this broke the patcher.

Unfortunately the patcher does use one class that's in studio:
PatchRunner.RestartRequiredException, so the solution is to use a new
class loader for the patcher that first tries to load classes from the
patcher jar and then from the parent loader (in this case, the loader
for the PatchRunner in studio).

Test: Added a test
Fixes: 268282800
Change-Id: I01453b62fa4ed3a818ecd321eaf2dc3f790f5bb0