Refactor annotation wrapper
This change removes the PsiWrapper package's annotation helper, opting
instead to actually create an "annotated" wrapper. The wrappers
representing elements that can have annotations will now use that as a
base class/interface, so that the logic therein can be shared more
easily.

Additionally, I've created a full "annotation" wrapper. This is to help
with upcoming changes were we'll need to examine the annotations in more
detail.

Bug: 265846405
Test: Included
Change-Id: I6f47a7e7ec27decdd180610438043e2ed9fb9b29