Fix live edit status after termination
Problem: Running the app on a second device after terminating the
app on another device (with Live Editing of the same app on) will
result in Live Edit being unresponsive on the second device.

Analysis: This problem occurs because there is a check to determine
if there are any disabled Live Edit devices, and if so, prevents the
Live Edit Deploy Monitor from processing changes. The error here
was checking for the existance of any such devices, rather than
ensuring all devices met this criteria.

Solution: The fix was simply to check all devices for Disabled
status, instead of returning on the first such occurence.

However, it is also apparent this bug was hiding a number of other
issues. So this commit fixes the newly discovered issues where
status updates were not checking if a device is already Disabled.

Bug: N/A
Test: manual
Change-Id: Ifb90c16ec716930b394c34185a38840e7b4aa6ec