[AUA] Tweak notification balloon implementation
The existing implementation (from the platform) has two codepaths.
The one we are using (before this change) generates some invisible
ActionLinks in the action panel in order to populate the drop-down.
This apparently leaves residual active areas where users can click on
apparently-blank space and an action will be taken.

Since what we are doing is a suggestion, make our notifications be
suggestions.  (And make the one referring to deprecated AGP versions
be an important suggestion).  This changes the codepath used in the
balloon implementation to one which does not use invisible
ActionLinks, but instead constructs a button (for the suggested
action) and a dropdown directly.  We have to re-order the actions in
order that the correct action (the first) is placed in the button.

Also add logging so that if something else goes wrong we stand a
chance of reconstructing what happened.

Bug: 269193617
Test: manual
Change-Id: Ice82a1d28cb9dea3d60ba3dedff067e05eabd16f