Do not try to attach when we don't know the version of compose
We were trying to connect with the compose inspector when we
received as version from the compat check without checking if the
version was incompatible.

With andrdoidx.compose.ui:ui versions:
- 1.4.0-alpha05
- 1.4.0-beta01

The version file is garbled. We will receive a status of
incompatible but we tried to connect anyway.

Now: show the incompatible error instead.
Also: update the error message since it is unlikely that the user
is using a version prior to 1.0.0-beta02.

Bug: 268730525
Test: Updated unit test
Change-Id: I9801f3f23d7774bed936a32a3bf25fa5f236ee47