Remove unused LambdaPropertyItem class.
This class is redundant with LambdaParameterItem, which better handles
threading concerns. The LambdaPropertyItem class is not used anywhere
besides its own test and a single other test, which can easily be
substituted to use LambdaParameterItem instead.

This is in preparation for updates to LambdaResolver to handle the K2
plugin's analysis API - see go/studio-k2-migration for background
information on this migration.

Bug: 269235880
Test: bazel test //tools/adt/idea/layout-inspector:all
Change-Id: Ia9503b1c194cb255f374a2b3e6dff0bb1a46bf8f