Choose the right process event for session
When profiler determines a session's support level, i.e.,
debuggable or profileable, it starts from the PID and
queries the database for the corresponding Common.Process
event. Existing implementation always looks at the last
event before the session starts.

It usually works. However, on API 29 (Q) and 30 (R), a
debuggable process have two events. One is from the ddmlib/
adblib (via "adb track-jdwp") which would report the process
as debuggable; the other is the result for the daemon's
DISCOVER_PROFILEABLE command which would report the process
as profileable. Sometimes, the profileable event arrives
later than the debuggable one. In this scenario, the
session would mistakenly treat the debuggable process as
a profileable-but-not-debuggable one.

The fix is to scan multiple associated Common.Process events
and find the one with the highest exposure level.

Bug: 269507519
Test: Adding StudioProfilersTest#testDebuggableProcessPresentedAsDebuggable
Change-Id: Ie0e964b40b99025d3bf1d9075ad5b5780e3dbea5