Move forceRefresh to the use the refresh flow
This makes the flow the only entry point to refresh, ensuring that
multiple refreshes can not run in parallel.

forceRefresh did require a way to know when the refresh had completed so
we have added that into requestRefresh. If given a CompletableDeferred,
it will be called when the refresh has completed.
This, in turn, required to add a way to batch multiple buffered requests
in a way that we can maintain all the listeners. For that, this change
adds conflateLatest. It works similarly to conflate but allowing to
merge the waiting element with the arriving one.

Bug: N/A
Test: CompletableDeferredUtilsTest, ConflatingFlowTest
Change-Id: I8007c15d1b260eb56a5910f6bd9d499392c3cbac