Handle attribute conflict in dynamic feature manifest merger
ManifestPanel currently parse the error message from manifest merger logging records to determine the position of error to generate a link to that file position. This does not work in some cases when the manifest merger could not determine the error position. The regex to parse the error message to determine the file position fails because the error message returned is different. This change only addresses the way we parse the error message to accomodate this specific error condition when the position of error cannot be determined by manifest merger.

Follow up work would be for manifest merger to  provide structured error messages instead of raw strings and to fix the way we compute the manifest files that contribute to a dynamic feature's manifest merger. Currently, all of app's dependencies are also included in the corresponding dynamic feature's manifest merger. Another problem is the spurious APPLICATION_ID(with value uninitialized_application_id) included as manifest override for dynamic features. This can result in unnecessary attribute value conflicts in case of an app's dependency with app id placeholder in manifest file.

Bug: 269085620
Test: ManifestConflictTest(added two new cases)
Change-Id: Id1613618c3d113985a21af8c3856b8a8574e1675
(cherry picked from commit 18bfbbe0e168511cfc2a8f972ea4cbb9658a644e)