[toml] Put banners and other caveats behind a flag
Preserve existing behavior for now, but turning the all the IDE-based
Version Catalog notifications off should now be as simple as switching
a single boolean.

Switching the boolean does not have quite the same effect as deleting
all related code: detecting whether the project uses Version Catalogs
at all has at least the side-effect of logging that fact.  When
removing the flag we will need to evaluate whether preserving
detection and logging on projects is important.

Bug: 267644534
Test: existing
Change-Id: Iad5d8a26b8da7ee305c8399db2182f74f5ec132b